Major refactor of rewards - WIP: Builds, runs but badly broken
Split out reward calculations to only be based on snapshots, to
improve test surface and formalise dependencies.  Also will allow
this to run asynchronously later.
